:recycle: signUp 메소드 도메인 변경 User -> Auth
