+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/auth/dto/LoginRequest.kt
@@ -1,4 +1,4 @@
-package com.teamsparta.exhibitionnewsfeed.domain.user.dto
+package com.teamsparta.exhibitionnewsfeed.domain.auth.dto
 
 import jakarta.validation.constraints.Email
 import jakarta.validation.constraints.NotBlank
