+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/auth/dto/SignUpResponse.kt
@@ -1,4 +1,4 @@
-package com.teamsparta.exhibitionnewsfeed.domain.user.dto
+package com.teamsparta.exhibitionnewsfeed.domain.auth.dto
 
 import com.teamsparta.exhibitionnewsfeed.domain.user.model.User
