+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/auth/service/AuthServiceImpl.kt
@@ -3,10 +3,12 @@ package com.teamsparta.exhibitionnewsfeed.domain.auth.service
 import com.teamsparta.exhibitionnewsfeed.domain.auth.AuthUser
 import com.teamsparta.exhibitionnewsfeed.domain.auth.JwtTokenProvider
 import com.teamsparta.exhibitionnewsfeed.domain.auth.TokenType
+import com.teamsparta.exhibitionnewsfeed.domain.auth.dto.LoginRequest
+import com.teamsparta.exhibitionnewsfeed.domain.auth.dto.LoginResponse
+import com.teamsparta.exhibitionnewsfeed.domain.auth.dto.SignUpRequest
+import com.teamsparta.exhibitionnewsfeed.domain.auth.dto.SignUpResponse
 import com.teamsparta.exhibitionnewsfeed.domain.auth.model.RefreshToken
 import com.teamsparta.exhibitionnewsfeed.domain.auth.repository.RefreshTokenRepository
-import com.teamsparta.exhibitionnewsfeed.domain.user.dto.LoginRequest
-import com.teamsparta.exhibitionnewsfeed.domain.user.dto.LoginResponse
 import com.teamsparta.exhibitionnewsfeed.domain.user.repository.UserRepository
 import com.teamsparta.exhibitionnewsfeed.exception.ModelNotFoundException
 import com.teamsparta.exhibitionnewsfeed.exception.UnauthorizedException
@@ -22,6 +24,12 @@ class AuthServiceImpl(
     private val jwtTokenProvider: JwtTokenProvider,
     private val passwordEncoder: PasswordEncoder,
 ) : AuthService {
+
+    @Transactional
+    override fun signUp(request: SignUpRequest): SignUpResponse {
+        if (userRepository.existsByEmail(request.email)) throw IllegalStateException("이미 존재하는 회원입니다.")
+        return SignUpResponse.from(userRepository.save(request.toEntity(passwordEncoder)))
+    }
 
     @Transactional
     override fun login(request: LoginRequest): LoginResponse {
